API: refactoring (reorder common variables).
Variables like: `MC_STEPS`, `STEP_TYPES`, `DATE_FORMAT` -- are more
about common data for all API submodules than "configuration" or "common
data for storages"; so they were gathered in `api.common` and removed
from `api.storages.common`, `api.storages.es.common` and `api.config`.
