+++ Utils/API/server/lib/dkb/api/common.py
@@ -1,0 +1,24 @@
+"""
+Common variables and dictionaries for API submodules.
+"""
+
+# Processing step types
+STEP_TYPES = ['step', 'ctag_format']
+
+# MC campaign steps order
+MC_STEPS = ['Evgen',
+            'Evgen Merge',
+            'Simul',
+            'Simul(Fast)',
+            'Merge',
+            'Digi',
+            'Reco',
+            'Rec Merge',
+            'Atlfast',
+            'Atlf Merge',
+            'TAG',
+            'Deriv',
+            'Deriv Merge']
+
+# Default datetime format
+DATE_FORMAT = '%d-%m-%Y %H:%M:%S'
