--- Utils/API/server/lib/dkb/api/storages/common.py
@@ -1,6 +1,0 @@
-"""
-Common values for all storages.
-"""
-
-# Processing step types
-STEP_TYPES = ['step', 'ctag_format']
